Change the way that context is handled.
Now all the static declaration of monads and applicatives
are called `context` and them are identified by the
`ContextClass` protocol.

The protocol has an method that return the context priority.
The priority is a tool used for determine when some context
has precedence that other. As great example, monad-transformers
are most important in context stack that simple monads.

At this moment, only two values are used for context level: 10
and 100; the first for simple context and the last for
transformers. In future that values will be defined in some
kind of constants.